+++ CommHandler/CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [2.4.2] - 2025-09-09
+## [2.4.2] - 2025-09-24
 
 ### Fixed
 
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Updated specification to `2.1_2025-09-02`, adds W/m2
 - updated Gradle build file
+- query parameters and headers are not added to REST request if their value is empty
 
 ## [2.4.1] - 2025-05-20